more cleanup

+++ routes/recipe.js
@@ -36,6 +36,9 @@ const constructTags = (recipe) => {
 
 router.get('/:recipeId', (request, response, next) => {
     const { recipeId } = request.params;
+    if (!request.session.userId) {
+      return response.status(403).json({success: false, message: 'Access denied: No session for the user.'})
+    }
     let id = request.session.userId;
     client.query('SELECT * FROM recipes WHERE user_id=$1 AND id=$2',
     [id, recipeId],
@@ -53,48 +56,52 @@ router.get('/:recipeId', (request, response, next) => {
           directions: recipe.directions, 
           tags: constructTags(recipe)
         }
-        response.status(200).json(recipe_response)
+        response.status(200).json({ success: true, recipe: recipe_response })
       } else {
-        response.status(500).json(res)
+        response.status(500).json({ success: false, message: 'No recipe could be found.'})
       }
     });
   });
 
 router.put('/:recipeId', (request, response, next) => {
-    const { recipeId } = request.params;
-    const { category } = request.body;
-    let id = request.session.userId;
-    if (id) {
-      client.query('UPDATE recipes SET category=$1 WHERE id=$2',
-        [category, recipeId],
-        (err, res) => {
-          if (err) {
-            return next(err);
-          }
-          if (res) {
-            return response.status(200).json({success: true})
-          } else {
-            return response.status(500).json({success: false, message: 'could not update DB'})
-          }
-      })
-    } else {
-      return response.status(401).json({error: true, message: 'You are not authenticated.'})
-    }  
+  if (!request.session.userId) {
+    return response.status(403).json({success: false, message: 'Access denied: No session for the user.'})
+  }
+  const { recipeId } = request.params;
+  const { category } = request.body;
+  let id = request.session.userId;
+  if (id) {
+    client.query('UPDATE recipes SET category=$1 WHERE id=$2',
+      [category, recipeId],
+      (err, res) => {
+        if (err) return next(err)
+        if (res) {
+          return response.status(200).json({ success: true, message: 'Recipe updated.' })
+        } else {
+          return response.status(500).json({success: false, message: 'Could not update recipe.'})
+        }
+    })
+  } else {
+    return response.status(401).json({error: true, message: 'You are not authenticated.'})
+  }  
 });
 
 
 router.delete('/:recipeId', (request, response, next) => {
-    const { recipeId } = request.params;
-    client.query('DELETE FROM recipes WHERE id=$1',
-    [recipeId],
-        (err, res) => {
-        if (err) return next(err);
-        if (res) {
-        return response.status(200).json({success: true})
-        } else {
-        return response.status(500).json({success: false, message: 'could not delete from DB'})
-        }
-    });
+  if (!request.session.userId) {
+    return response.status(403).json({success: false, message: 'Access denied: No session for the user.'})
+  }
+  const { recipeId } = request.params;
+  client.query('DELETE FROM recipes WHERE id=$1',
+  [recipeId],
+      (err, res) => {
+      if (err) return next(err);
+      if (res) {
+        return response.status(200).json({ success: true, message: 'Recipe deleted.' })
+      } else {
+        return response.status(500).json({success: false, message: 'Recipe not deleted.'})
+      }
+  });
 });
 
 module.exports = router;
